src/webhook_handler/pyproject.toml
[tool.poetry]
name = "hunters-incoming"
version = "0.1.0"
description = ""
authors = ["Ross Brown <ross@crema.us>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
aiokafka = "^0.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
